Organization and new StartMenu
Organized some of the scripts.
Created a new StartMenu scene for the main menu of the game. Currently
it only gives the option to go to the AStar test or the AI Trainer.

Moved over the new PauseMenu scripts from SEE and implemented the
temporary pause menu for the StartMenu.