Preparations
Created TestableAgent class which will be the base for all agents using
an ANN.
Created Human class. Has not been tested.
Created generic Genome class to have variations in different learning
and acting ANNs.
Created the BasicGenome which will be a copy of the genome and genetic
algorithm code from Assignment 2.

Updated the DebugRenderer to use the current camera instead of
Camera.main.
Never use Camera.main since that will only be the first ever camera
used in the project/build. Instead use DebugRenderer.currentCamera.
DebugRenderer.currentCamera is updated in the CameraController class.
All scenes now have an instance of this class, but there is an option
to freeze the camera.

Updated all scripts to use new currentCamera.

Updated DebugRenderer to allow colors for drawBox.
Updated feelers to be a transparent red.
Updated pieSlices to be a transparent blue.
(Transparent colors to see the overlap)