TestScene, MapEditor, and other things
Agents cannot move outside the map/grid (but still can be pushed).
Properly get the heading of agents in the start method.
Static createAgent method that will instantiate an agent and set the
proper values.

DebugRenderer now allows for colored circles.

Camera position now stops moving when the map edges get close to the
center of the screen.

Created an Options class where we can store different options. Probably
going to be renamed to Settings or GameData later. As of now, only
stores the name of the map that should be loaded and the GameMap (see
below) for that map once loaded.

Created a format for maps that include human spawn points, penguin
spawn points, and the ICE machine location.
Updated map to be able to load and save this new format.

Created a class GameMap to be the manager of the map and grid related
items. (Will eventually replace Background.cs) Grid will now contain
the same number of cells as the map.

Created a simple map editor for the new format.
Left click to place item. Press 1 - 4 to change items.
Color in the top left indicates the current selected item.

Black means walls
Red means human spawn (can only be placed on the edges of the map)
Blue means penguin spawn (only one)
Cyan means Ice machine location (only one)

Press F2 to save, but currently it just saves as the current map name
with New appended.

Created a test scene to test the maps. Right clicking spawns a human at
a random spawn point.

Updated humans, testable agents, genomes, and neural network to allow
for creation and testing.

Added buttons to the temporary start menu to access map editor and
tests for human ANNs.