Setup Genetic Algorithm for testing
Lots of things so that new ANNs are easier to train.

Created a new Genetic Algorithm class that uses the new generic
Genomes. Lots of updates in order to get that working. Also updated the
start menu to give a gui way of starting a new population, or loading a
previous for testing.

Agents are now forced to stay within bounds rather than just not being
able to move outside the bounds.

Deleted all previous old population files.

Press F3 to freeze or unfreeze the camera. The camera movements will
need a lot of work to be more user friendly in the final game.