Play, MapEditor, AI Training updates
Created a scene for playing the actual game. Implemented a wave manager
for when to send out units.

Updated MapEditor to register clicks on each frame.

Updated the AI Training scene so that you can load a map on initial
population creation. Created many different genomes and tested some.
Entire population is tested at the same time instead of simply testing
one individual at a time.

Added drawLineRect method in DebugRenderer to draw a rectangle with a
specific thickness for the lines.

Created a minimap to display the map and all agents on the map. Right
click dragging will move the minimap. Left clicking will move the
camera to the click location.

Moved some button checks from FixedUpdate to Update.