Human flocking behavior and boundaries
Added boundaries to the game map so that humans will not try to go to
edges of the map thinking there is a path.

Updated human behavior for when the ICE Machine is being held. The
holder of the ICE Machine now moves slower (0.6 times slower). Other
humans will gather around the holder, but avoid getting in front and
getting too close.

Fixed a bug on scene end where humans would try to access the ICE
Machine after it had already been deconstructed.

Debug circles are now only displayed in the MapEditor.
Right click in the MapEditor to remove item.