Penguins are now GameAgents
Added the abstract checkButtons method in agent so that buttons presses
are checked in update rather than fixed update.
Added the abstract drawStatus method in agent so that GUI information
can be drawn separate from debug information. (Ex. health bar)

Converted IciclePenguin to GameAgent.
Removed all transitions to sleep state and instead placed the
transition in the onDeath method that is called when health is below 0.

Todo: move HumanAgent to GameAgent