Start of PlayGame GUI and Camera Controller tweaks
Camera controller now only moves when the mouse is within the 100 pixel
edge for the viewport. Max zoom is now automatically calculated based
on map size.

Created a simple GUI display for the PlayGame scene to include current
wave, time till next wave, time taken on current wave, number of humans
killed, number of humans on the map, number of penguins on the map,
number of penguins sleeping (todo), a button to spawn penguins, and the
minimap.

Wall size was doubled.  Agent speeds may need to be adjusted.

Added a minimap to MapEditor. Spawn points are shown instead of agent
locations.
Added a scroll bar to minimap to change zoom level.

Removed old ice machine sprites