Humans now attack penguins
Human checks for nearest unobstructed awake penguin and targets that
penguin.
If no target found, human is in move state.

Humans seem to turn really quickly though.

Note, if penguin is in chillin state and they are hit with a
tranquilizer, they should probably attack. Unless we want that to be a
feature? Not knowing if they were hit or not. (Make Adjacent Agent
sensor the same size as humans to give them vision)
Had to modify the penguin class to make the fsm visible to know if it
was sleeping or not. Maybe change to a boolean check later?

Tweaked the tranquilizer values so that they are slightly slower, deal
less damage, and damage is spread out over more ticks.

Adjacent Agents sensor can now take in a type of agent to look for. If
nothing is passed in the constructor, it will search for all agents. No
check is done to see if the type passed in is a child of Agent.