Testing sprites
Created a sprite for the Icicle Club and a GameObject to go with it. I
suggest adding a monobehavior script to go with it for actually dealing
the damage but I’m welcome to hearing your ideas on how you want to do
it too.

Organized the Sprites folder with more folders.
Attempted to make new wall sprites but don’t like how they turned out.

Updated Map to draw walls to their set size rather than the renderer’s
size.

Updated CameraController to stop the null pointer error in AStar.unity

Updated Tranquilizers to ignore humans since that bugged me too much.
May need to be tweaked later if it becomes noticeable that they move
through human agents.