High Scores
Scores and HighScores list created. Scores include the wave reached,
the number of humans killed, the total time played (might need to
include/remove time skipped when starting a wave early), and the time
of score creation. Scores are saved into the HighScores list saved in
the Options class. On load of the game, the HighScores file is read.
Whenever a new score is created, the HighScores file is rewritten. Only
the top 10 scores are saved.

Changed some formatting for displayed values.
Added code for spawning initial penguins. Commented out for now.
Fixed the MapEditor from registering clicks when using the MiniMap.