GUI Updates
All components of the GUI are now drawn to the screen with the minimum
screen width of 800 pixels. All GUI components are drawn in the bottom
center.

Fixed Options not loading the file before loading the HighScores.

Limit the number of penguin spawning to 4 at the start and one
additional penguin per wave completed.

Made small map refined.

Scores now save the map.