Start of Dynamic NavGraph
Penguins will move around other agents on the map.
May lag on larger maps. (Due to minimap?)

Made navgraph twice as large to make more nodes. When an agent is on
top of a node, the node is taken and not considered by AStar.

Mini map now displays penguins as green and sleeping penguins as blue