Map Editor GUI
Created a GUI for the Map Editor displaying placable items.
Clicking on an item in the GUI switches to that item.
A white box is drawn around the current item.

Updated the map editor pause menu to be more centered.

Fixed bug when creating new maps. Forgot to update default locations to
the new cell format.